[refactor] fixture 분리 (#416)
