+++ backend/src/test/java/dev/tripdraw/draw/application/TripUpdateEventHandlerIntegrationTest.java
@@ -1,6 +1,6 @@
 package dev.tripdraw.draw.application;
 
-import static dev.tripdraw.test.fixture.TestFixture.여행;
+import static dev.tripdraw.test.fixture.TripFixture.여행;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
